move source coverage collection to separate class
