+++ lib/remap.js
@@ -14,27 +14,7 @@ const updateBranches = require('./updateBranches');
 
 const sourceMapRegEx = /(?:\/{2}[#@]{1,2}|\/\*)\s+sourceMappingURL\s*=\s*(data:(?:[^;]+;)+base64,)?(\S+)/;
 
-const MaybeWeakMap = typeof WeakMap === 'function' ? WeakMap : function () {
-  const objectList = [];
-  const dataList = [];
-
-  this.set = function (object, data) {
-    const index = objectList.indexOf(object);
-    if (index !== -1) {
-      data[index] = data;
-    } else {
-      objectList.push(object);
-      dataList.push(data);
-    }
-  };
-
-  this.get = function (object) {
-    const index = objectList.indexOf(object);
-    if (index !== -1) {
-      return dataList[index];
-    }
-  };
-};
+const SparceCoverageCollector = require('./SparceCoverageCollector');
 
 const getMapping = require('./getMapping');
 /**
@@ -96,43 +76,7 @@ function remap(coverage, options = {}) {
     coverage = [coverage];
   }
 
-  let srcCoverage = {};
-
-  const metaInfo = new MaybeWeakMap();
-
-  /**
-   * Generate a coverage object that will be filled with the remapped data
-   * @param  {string} filename    The name of the file that is being remapped
-   * @return {Object}             An object that provides the actual data and
-   *                              its shadow data used for reference.
-   */
-  const getSourceCoverage = (filename) => {
-    let data = srcCoverage[filename];
-    if (!data) {
-      data = srcCoverage[filename] = {
-        path: filename,
-        statementMap: {},
-        fnMap: {},
-        branchMap: {},
-        s: {},
-        b: {},
-        f: {},
-      };
-      metaInfo.set(data, {
-        indexes: {},
-        lastIndex: {
-          s: 0,
-          b: 0,
-          f: 0,
-        },
-      });
-    }
-
-    return {
-      data,
-      meta: metaInfo.get(data),
-    };
-  };
+  const sparceCoverageCollector = new SparceCoverageCollector();
 
   coverage.forEach((item) => {
     Object.keys(item).forEach((filePath) => {
@@ -183,7 +127,7 @@ function remap(coverage, options = {}) {
         } catch (error) {
           warn(new Error(`Could find source for : "${filePath}"`));
         }
-        srcCoverage[filePath] = fileCoverage;
+        sparceCoverageCollector.setCoverage(filePath, fileCoverage);
         return;
       }
 
@@ -218,7 +162,7 @@ function remap(coverage, options = {}) {
 
         sourceMap.sourcesContent.forEach((source, idx) => {
           inlineSourceMap[sourceMap.sources[idx]] = true;
-          getSourceCoverage(sourceMap.sources[idx]).data.code = codeIsArray ? source.split('\n') : source;
+          sparceCoverageCollector.setSourceCode(sourceMap.sources[idx], codeIsArray ? source.split('\n') : source);
           if (sourceStore) {
             sourceStore.set(sourceMap.sources[idx], source);
           }
@@ -241,9 +185,12 @@ function remap(coverage, options = {}) {
         return Object.assign(mapping, { source: resolvePath(mapping.source) });
       };
 
-      updateFunctions(fileCoverage, getMappingResolved, getSourceCoverage);
-      updateStatements(fileCoverage, getMappingResolved, getSourceCoverage);
-      updateBranches(fileCoverage, getMappingResolved, getSourceCoverage);
+      updateFunctions(fileCoverage, getMappingResolved, sparceCoverageCollector);
+      updateStatements(fileCoverage, getMappingResolved, sparceCoverageCollector);
+      updateBranches(fileCoverage, getMappingResolved, sparceCoverageCollector);
+
+      // todo: refactor exposing implementation details
+      const srcCoverage = sparceCoverageCollector.getFilesCoverage();
 
       if (sourceMap.sourcesContent && options.basePath) {
         // Convert path to use base path option
@@ -266,14 +213,14 @@ function remap(coverage, options = {}) {
 
   const collector = new Collector();
 
-  srcCoverage = Object.keys(srcCoverage)
+  const srcCoverage = sparceCoverageCollector.getFilesCoverage();
+
+  collector.add(Object.keys(srcCoverage)
     .filter((filePath) => !(exclude && exclude(filePath)))
     .reduce((obj, name) => {
       obj[name] = srcCoverage[name];
       return obj;
-    }, {});
-
-  collector.add(srcCoverage);
+    }, {}));
 
   /* refreshes the line counts for reports */
   collector.getFinalCoverage();
