+++ lib/updateStatements.js
@@ -1,24 +1,4 @@
-const updateFunctions = (fileCoverage, getMappingResolved, getSourceCoverage) => {
-  function updateItem(source, srcItem, hits) {
-    const { data, meta } = getSourceCoverage(source);
-
-    const key = [
-      's',
-      srcItem.start.line, srcItem.start.column,
-      srcItem.end.line, srcItem.end.column,
-    ].join(':');
-
-    let index = meta.indexes[key];
-    if (!index) {
-      index = ++meta.lastIndex.s;
-      meta.indexes[key] = index;
-      data.statementMap[index] = srcItem;
-    }
-
-    data.s[index] = data.s[index] || 0;
-    data.s[index] += hits;
-  }
-
+const updateFunctions = (fileCoverage, getMappingResolved, sparceCoverageCollector) => {
   Object.keys(fileCoverage.statementMap).forEach((index) => {
     const genItem = fileCoverage.statementMap[index];
     const hits = fileCoverage.s[index];
@@ -26,7 +6,7 @@ const updateFunctions = (fileCoverage, getMappingResolved, getSourceCoverage) =>
     const mapping = getMappingResolved(genItem);
 
     if (mapping) {
-      updateItem(mapping.source, mapping.loc, hits);
+      sparceCoverageCollector.updateStatement(mapping.source, mapping.loc, hits);
     }
   });
 };
